pyproject.toml
[tool.poetry]
name = "reloadradar"
version = "0.1.0"
description = ""
authors = ["Kenan <kenan.altaki@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11.4"
Django = "^4.2.6"
requests-html = "^0.10.0"
pandas = "^2.1.1"
djangorestframework = "^3.14.0"
markdown = "^3.5"
django-filter = "^23.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
